cwod_site/requirements.txt
BeautifulSoup==3.2.0
Django==1.3
MySQL-python==1.2.3
PyYAML==3.09
argparse==1.2.1
boto==2.0
distribute==0.6.10
django-debug-toolbar==0.8.5
-e git://github.com/sunlightlabs/django-locksmith.git#egg=django_locksmith-0.5.0
django-mediasync==2.2.0
django-piston==0.2.2
django-typogrify==1.2.2
gunicorn==0.11.2
httplib2==0.6.0
ipython==0.10.1
lxml==2.3beta1
mimeparse==0.1.3
name-tools==0.1.2
nltk==2.0b9
numpy==1.5.1
pygooglechart==0.3.0
pymongo==1.9
python-dateutil==1.5
python-sunlightapi==1.0.0
